Extract finalized redaction helpers for component review
